refact: create common to enums
