+++ src/common/enums/tipo-.conta.enum.ts
@@ -1,0 +1,18 @@
+export enum TipoConta {
+  CORRENTE = 'CORRENTE',
+  POUPANCA = 'POUPANCA',
+}
+
+export enum TipoCargo {
+  AGENTE = 'Agente',
+  GERENTE = 'Gerente',
+}
+
+export enum TipoTransacao {
+  DEBITO = 'debito',
+  CREDITO = 'credito',
+  DEPOSITO = "DEPOSITO",
+  SAQUE = "SAQUE",
+  TRANSFERENCIA = "TRANSFERENCIA",
+  PAGAMENTO = "PAGAMENTO",
+}
