+++ src/contas/entities/conta-corrente.entity.ts
@@ -1,5 +1,6 @@
 import { Transacao } from "src/transacoes/entities/transacao.entity";
 import { Contas } from "./conta.entity";
+import { TipoConta } from "src/common/enums/tipo-.conta.enum";
 
 
 export class ContaCorrente extends Contas {
@@ -8,8 +9,9 @@ export class ContaCorrente extends Contas {
     numeroConta: number,
     saldo: number,
     transacoes: Transacao[],
+    tipoConta: TipoConta.CORRENTE,
     public limiteChequeEspecial: number,
   ) {
-    super(id, numeroConta, saldo, transacoes);
+    super(id, numeroConta, saldo, transacoes,tipoConta);
   }
 }
