+++ src/contas/entities/conta-poupanca.entity.ts
@@ -1,5 +1,6 @@
 import { Transacao } from 'src/transacoes/entities/transacao.entity';
 import { Contas } from './conta.entity';
+import { TipoConta } from 'src/common/enums/tipo-.conta.enum';
 
 const TAXA_JUROS_MENSAL = 0.5 / 100;
 
@@ -10,10 +11,11 @@ export class ContaPoupanca extends Contas {
     numeroConta: number,
     saldo: number,
     transacoes: Transacao[],
+    tipoConta: TipoConta.POUPANCA,
     public taxaJuros: number = TAXA_JUROS_MENSAL,
     ultimoCalculoJuros?: Date,
   ) {
-    super(id, numeroConta, saldo, transacoes);
+    super(id, numeroConta, saldo, transacoes, tipoConta);
     this.ultimoCalculoJuros = ultimoCalculoJuros;
   }
 }
